refactor: remove unused classes and methods
